pyproject.toml
[build-system]
requires = [
  "setuptools>=61.0",
  "wheel"
]
build-backend = "setuptools.build_meta"

[project]
name = "permissible"
version = "0.2.17"
description = "Extended, flexible and powerful object-level and rule-driven permissions for Django & Django REST framework"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
license-files = ["LICEN[CS]E*"]
dependencies = [
  "Django>=5,<6",
  "djangorestframework>=3",
  "django-guardian",
  "djangorestframework-guardian"
]

[project.urls]
Homepage = "https://github.com/gaussian/permissible"
Issues = "https://github.com/gaussian/permissible/issues"

[tool.setuptools]
packages = ["find:"]
